requirements.txt
boto3==1.16.47
dj-database-url==0.5.0
django==3.1.4
django-configurations==2.2
django-filter==2.4.0
django-public-admin==0.0.5
django-simple-history==2.12.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dramatiq[rabbitmq,watch]==1.10.0
gunicorn==20.0.4
jinja2==2.11.2
psycopg2-binary==2.8.6
python-dotenv==0.15.0
schematics==2.1.0
scrapy==2.4.1
sentry-dramatiq==0.3.2
sentry-sdk==0.19.5
spidermon==1.14.0
tika==1.24
uritemplate==3.0.1
whitenoise==5.2.0
xlrd==2.0.1
